# api/openapi.yaml
openapi: 3.0.3
info:
  title: DeepProof Nexus API
  version: 1.0.0
  description: Decentralized Deepfake Detection & Perceptual Provenance Network on 0G
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.deepproof.ai/api/v1
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /detect:
    post:
      summary: Submit file for deepfake detection
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                recipientAddress:
                  type: string
                  description: Ethereum address to receive SPT (optional)
      responses:
        '202':
          description: Detection job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string }
                  sha256: { type: string }
                  status: { type: string, enum: [queued] }
  /detect/{jobId}:
    get:
      summary: Poll detection job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [queued, detecting, uploading_evidence, anchoring, complete, error] }
                  result: { type: object }
                  registryTx: { type: string }
                  mintTx: { type: string }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key